# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sales Server",
    "description": "Documentation API. https://github.com/danielschmitz/sales-server <br/><br/>",
    "version": "1.0",
    "contact": {
      "name": "Daniel Schmitz",
      "email": "danieljfa@gmail.com",
      "url": "https://github.com/danielschmitz"
    }
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header"
    }
  },
  "paths": {
...
    "/category/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get a Category",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Category",
            "schema": {
              "$ref": "#/definitions/CategoryResult"
            }
          },
          "404": {
            "description": "Category not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "🔒️ Edit a Category",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "category",
            "in": "body",
            "description": "Category Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category registered successfully.",
            "schema": {
              "$ref": "#/definitions/CategoryResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a category with that name"
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "🔒️ Delete a Category",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          },
          "404": {
            "description": "Category not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/category": {
      "post": {
        "tags": ["Categories"],
        "summary": "🔒️ Create a new Category",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "category",
            "in": "body",
            "description": "Category Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category registered successfully.",
            "schema": {
              "$ref": "#/definitions/CategoryResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a category with that name"
          }
        }
      }
    },
...
    "/shipper/{id}": {
      "get": {
        "tags": ["Shippers"],
        "summary": "Get a Shipper",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Shipper",
            "schema": {
              "$ref": "#/definitions/ShipperResult"
            }
          },
          "404": {
            "description": "Shipper not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["Shippers"],
        "summary": "🔒️ Edit a Shipper",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "shipper",
            "in": "body",
            "description": "Shipper Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Shipper"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shipper registered successfully.",
            "schema": {
              "$ref": "#/definitions/ShipperResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a shipper with that name"
          }
        }
      },
      "delete": {
        "tags": ["Shippers"],
        "summary": "🔒️ Delete a Shipper",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Shipper deleted"
          },
          "404": {
            "description": "Shipper not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/shipper": {
      "post": {
        "tags": ["Shippers"],
        "summary": "🔒️ Create a new Shipper",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "shipper",
            "in": "body",
            "description": "Shipper Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Shipper"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shipper registered successfully.",
            "schema": {
              "$ref": "#/definitions/ShipperResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a shipper with that name"
          }
        }
      }
    },
...
    "/product/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get a Product",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "schema": {
              "$ref": "#/definitions/ProductResult"
            }
          },
          "404": {
            "description": "Product not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "🔒️ Edit a Product",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "product",
            "in": "body",
            "description": "Product Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product registered successfully.",
            "schema": {
              "$ref": "#/definitions/ProductResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a product with that name"
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "🔒️ Delete a Product",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          },
          "404": {
            "description": "Product not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/product": {
      "post": {
        "tags": ["Products"],
        "summary": "🔒️ Create a new Product",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "product",
            "in": "body",
            "description": "Product Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product registered successfully.",
            "schema": {
              "$ref": "#/definitions/ProductResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a product with that name"
          }
        }
      }
    }
  },
  "definitions": {
    "CategoryResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Category Name"
        },
        "description": {
          "type": "string",
          "example": "Category Description"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Category Name"
        },
        "description": {
          "type": "string",
          "example": "Category Description"
        }
      }
    },
...
    "ShipperResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "companyName": {
          "type": "string",
          "example": "Shipper Company Name"
        },
        "contactName": {
          "type": "string",
          "example": "Shipper Contact Name"
        },
        "contactEmail": {
          "type": "string",
          "example": "supplier@email.com"
        },
        "address_id": {
          "type": "number",
          "example": 10
        },
        "address": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 10
            },
            "street": {
              "type": "string",
              "example": "Street Address"
            },
            "city": {
              "type": "string",
              "example": "City"
            },
            "region": {
              "type": "string",
              "example": "Region"
            },
            "postalCode": {
              "type": "string",
              "example": "Postal Code"
            },
            "country": {
              "type": "string",
              "example": "Country"
            },
            "phone": {
              "type": "string",
              "example": "Phone"
            }
          }
        }
      }
    },
...
    "Shipper": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string",
          "example": "Shipper Company Name"
        },
        "contactName": {
          "type": "string",
          "example": "Shipper Contact Name"
        },
        "contactEmail": {
          "type": "string",
          "example": "supplier@email.com"
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string",
              "example": "Street Address"
            },
            "city": {
              "type": "string",
              "example": "City"
            },
            "region": {
              "type": "string",
              "example": "Region"
            },
            "postalCode": {
              "type": "string",
              "example": "Postal Code"
            },
            "country": {
              "type": "string",
              "example": "Country"
            },
            "phone": {
              "type": "string",
              "example": "Phone"
            }
          }
        }
      }
    },
    "ProductResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Chef Anton's Cajun Seasoning"
        },
        "quantityPerUnit": {
          "type": "string",
          "example": "48 - 6 oz jars"
        },
        "unitsInStock": {
          "type": "number",
          "example": 53
        },
        "unitPrice": {
          "type": "number",
          "example": 22.1
        },
        "supplier_id": {
          "type": "number",
          "example": 1
        },
        "category_id": {
          "type": "number",
          "example": 1
        },
        "supplier": {
          "type": "object",
          "properties": {
            "companyName": {
              "type": "string",
              "example": "Supplier Company Name"
            },
            "contactName": {
              "type": "string",
              "example": "Supplier Contact Name"
            },
            "contactEmail": {
              "type": "string",
              "example": "supplier@email.com"
            },
            "address_id": {
              "type": "number",
              "example": 10
            },
            "address": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 10
                },
                "street": {
                  "type": "string",
                  "example": "Street Address"
                },
                "city": {
                  "type": "string",
                  "example": "City"
                },
                "region": {
                  "type": "string",
                  "example": "Region"
                },
                "postalCode": {
                  "type": "string",
                  "example": "Postal Code"
                },
                "country": {
                  "type": "string",
                  "example": "Country"
                },
                "phone": {
                  "type": "string",
                  "example": "Phone"
                }
              }
            }
          }
        },
        "category": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Category Name"
            },
            "description": {
              "type": "string",
              "example": "Category Description"
            }
          }
        }
      }
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Chef Anton's Cajun Seasoning"
        },
        "quantityPerUnit": {
          "type": "string",
          "example": "48 - 6 oz jars"
        },
        "unitsInStock": {
          "type": "number",
          "example": 53
        },
        "unitPrice": {
          "type": "number",
          "example": 22.1
        },
        "supplier_id": {
          "type": "number",
          "example": 1
        },
        "category_id": {
          "type": "number",
          "example": 1
        },
        "supplier": {
          "type": "object",
          "properties": {
            "companyName": {
              "type": "string",
              "example": "Supplier Company Name"
            },
            "contactName": {
              "type": "string",
              "example": "Supplier Contact Name"
            },
            "contactEmail": {
              "type": "string",
              "example": "supplier@email.com"
            },
            "address_id": {
              "type": "number",
              "example": 10
            },
            "address": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 10
                },
                "street": {
                  "type": "string",
                  "example": "Street Address"
                },
                "city": {
                  "type": "string",
                  "example": "City"
                },
                "region": {
                  "type": "string",
                  "example": "Region"
                },
                "postalCode": {
                  "type": "string",
                  "example": "Postal Code"
                },
                "country": {
                  "type": "string",
                  "example": "Country"
                },
                "phone": {
                  "type": "string",
                  "example": "Phone"
                }
              }
            }
          }
        },
        "category": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Category Name"
            },
            "description": {
              "type": "string",
              "example": "Category Description"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}